Addresses adding fixed ips by way of Nova
After a recent patch to improve retry logic, the signature changed for
allocate_ip_address. Tests that inadvertently called allocate_ip_address
by way of testing update_port continued to work because
allocate_ip_address is stubbed, as it should be. However, a new test
introduced in this patch manually mocks the method, attempting to a list
in the signature for the method, which should catch many cases. The only
real solution is an integration test to vet the functionality.